services/metric-service/etl/pyproject.toml
[project]
name = "metric-service-etl"
version = "0.1.0"
description = "metric-service-etl microservice"
requires-python = ">=3.12,<4.0"
dependencies = [
    "six>=1.17.0,<2.0.0",
    "clickhouse-driver>=0.2.9,<0.3",
    "kafka-python>=2.2.10,<2.3",
    "python-dotenv>=1.1.0,<2.0.0",
    "pydantic-settings>=2.8.1,<3.0.0",
    "aiohttp (>=3.11.18,<4.0.0)"
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
